requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==12.0.0
celery==5.5.3
redis==7.0.1

